treewide: massive refactoring for v0.2.0
The focus is now on working on the waveform (after applying a
low-pass filter). I created multiple simple basic building blocks
that I thoroughly test on real world audio samples. This commit is
the first one where the code reliably detects the two envelopes
(beats) in the "sample1_lowpassed--double-beat.wav" file.

The next step is to ensure that the following samples work:
- holiday_lowpassed--single-beat.wav
- sample1_lowpassed--long.wav
- holiday_lowpassed--long.wav